// backend/src/generated-api/models/create-survivor-request-body.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Survivor Fantasy League - OpenAPI 3.0.3
 * This is the Survivor Fantasy League Spec for API's
 *
 * OpenAPI spec version: 1.0.11
 * Contact: pklein111697@gmail.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

 /**
 * 
 *
 * @export
 * @interface CreateSurvivorRequestBody
 */
export interface CreateSurvivorRequestBody {

    /**
     * The season tied with this survivor's details
     *
     * @type {number}
     * @memberof CreateSurvivorRequestBody
     * @example 47
     */
    seasonId: number;

    /**
     * The survivor's first name
     *
     * @type {string}
     * @memberof CreateSurvivorRequestBody
     * @example Andy
     */
    firstName: string;

    /**
     * The survivor's last name
     *
     * @type {string}
     * @memberof CreateSurvivorRequestBody
     * @example Rueda
     */
    lastName: string;

    /**
     * The survivor's nickname
     *
     * @type {string}
     * @memberof CreateSurvivorRequestBody
     * @example The Brain
     */
    nickname: string;

    /**
     * The survivor's job
     *
     * @type {string}
     * @memberof CreateSurvivorRequestBody
     * @example Software Engineer
     */
    job: string;

    /**
     * The survivor's description
     *
     * @type {string}
     * @memberof CreateSurvivorRequestBody
     * @example A software engineer who loves to code
     */
    description: string;

    /**
     * The survivor's city of origin
     *
     * @type {string}
     * @memberof CreateSurvivorRequestBody
     * @example San Francisco
     */
    fromCity?: string;

    /**
     * The survivor's state of origin
     *
     * @type {string}
     * @memberof CreateSurvivorRequestBody
     * @example California
     */
    fromState: string;

    /**
     * The survivor's country of origin
     *
     * @type {string}
     * @memberof CreateSurvivorRequestBody
     * @example United States
     */
    fromCountry: string;

    /**
     * The survivor's age
     *
     * @type {number}
     * @memberof CreateSurvivorRequestBody
     * @example 25
     */
    age: number;
}